requirements.txt
fastapi
uvicorn
langchain
langchain-community
pydantic
slowapi
python-dotenv
openai
httpx
gunicorn